+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * pool: take event reference in `send_event` methods ([Yuki Kishimoto])
 * pool: use the relay ingester to perform actions ([Yuki Kishimoto])
 * pool: avoid spawning a task for every authentication request ([Yuki Kishimoto])
+* pool: use `std::sync::OnceLock` instead of `tokio::sync::OnceCell` ([Yuki Kishimoto])
 * lmdb: bump MSRV to 1.72.0 ([Yuki Kishimoto])
 * lmdb: implement event ingester ([Yuki Kishimoto])
 * lmdb: avoid spawning thread for read methods ([Yuki Kishimoto])